lab-4-docker-debug-pascal-maker/backend/requirements.txt
fastapi>=0.115.2
uvicorn>=0.31.1
psycopg2-binary>=2.9.9
python-multipart>=0.0.9